# docs/swagger.yaml
definitions:
  be-dashboard-nba_internal_presentation_presenter.Pagination:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.Pagination'
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_auth.UserResponse:
    properties:
      active:
        type: boolean
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      img_name:
        type: string
      img_path:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter_menu.CreateMenuRequest:
    properties:
      active:
        type: boolean
      description:
        maxLength: 100
        type: string
      display:
        type: boolean
      group:
        maxLength: 50
        minLength: 1
        type: string
      icon:
        maxLength: 50
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      parent_id:
        minimum: 1
        type: integer
      url:
        maxLength: 100
        type: string
    required:
    - active
    - display
    - group
    - name
    - url
    type: object
  be-dashboard-nba_internal_presentation_presenter_menu.MenuListItem:
    properties:
      children:
        items:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.MenuListItem'
        type: array
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
      url:
        type: string
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_menu.ReadMenuListResponse:
    properties:
      group_childs:
        items:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.MenuListItem'
        type: array
      group_name:
        type: string
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_menu.UpdateMenuRequest:
    properties:
      active:
        type: boolean
      description:
        maxLength: 100
        type: string
      display:
        type: boolean
      group:
        maxLength: 50
        minLength: 1
        type: string
      icon:
        maxLength: 50
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
      parent_id:
        minimum: 1
        type: integer
      url:
        maxLength: 100
        type: string
    required:
    - active
    - display
    - group
    - name
    type: object
  be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.CreateMenuPermissionRequest:
    properties:
      action_name:
        maxLength: 50
        minLength: 1
        type: string
      code:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - action_name
    - code
    type: object
  be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.ReadMenuPermissionResponse:
    properties:
      action_name:
        type: string
      code:
        type: string
      id:
        type: integer
      menu_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: Template API for Golang project
  title: Base APP Go
  version: "1.0"
paths:
...
  /auth/me:
    get:
      description: Fetches the profile details of the currently authenticated user.
        Requires Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_auth.UserResponse'
              type: object
        "401":
          description: Unauthorized (Invalid or missing token)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: User profile not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Current User Profile
      tags:
      - Authentication
  /menus:
    get:
      description: Retrieves a paginated and searchable list of all menus in the system.
      parameters:
      - description: Search term for name, description, or URL
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get menu list
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.ReadMenuListResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed get list menu
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      summary: Get List of All Menus (Admin)
      tags:
      - Menu
    post:
      consumes:
      - application/json
      description: Adds a new menu item to the system. Requires Bearer token.
      parameters:
      - description: Menu data to create
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully create menu
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid JSON or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Create Menu
      tags:
      - Menu
  /menus/{menu_id}:
    delete:
      consumes:
      - application/json
      description: Updates an existing menu item by its ID. Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete menu
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid Menu ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Update Menu
      tags:
      - Menu
    put:
      consumes:
      - application/json
      description: Updates an existing menu item by its ID. Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Menu data to update
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update menu
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid Menu ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Update Menu
      tags:
      - Menu
  /menus/{menu_id}/permissions:
    get:
      description: Retrieves a paginated and searchable list of permissions (e.g.,
        read, create) for a specific menu ID. Requires Bearer token.
      parameters:
      - description: ID of the Menu
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Search term for action_name
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Sort order (e.g., action_name ASC)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get menu permission list
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.ReadMenuPermissionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request (Invalid Menu ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (Menu ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error (Query parsing or usecase failure)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Menu Permission List
      tags:
      - Menu Permission
    post:
      consumes:
      - application/json
      description: Adds a new permission (e.g., 'read', 'create') to a specific menu
        by its ID. Requires Bearer token.
      parameters:
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      - description: Permission data to create
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_menu_menu-permission.CreateMenuPermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully create menu permission
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid Menu ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Create Menu Permission
      tags:
      - Menu Permission
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your token}" into the field below.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"